Add requirements.txt
Note:  The awscli and boto3 packages are only needed in the local
environment to enable local testing.  The smartsheet-python-sdk package
must be installed in the project root along with its dependencies, and
all of these and (and the sync script) must be zipped and uploaded to
the AWS Lambda Function.